split eos_spiner.hpp into seperate rhoT and rhoSie, maintaining eos_spiner.hpp as a connection to both. + clangformat
