+++ singularity-eos/CMakeLists.txt
@@ -45,6 +45,8 @@ register_headers(
     eos/eos_mgusup.hpp
     eos/eos_powermg.hpp
     eos/eos_spiner.hpp
+    eos/eos_spiner_rho_sie.hpp
+    eos/eos_spiner_rho_temp.hpp
     eos/eos_davis.hpp
     eos/eos_gruneisen.hpp
     eos/eos_vinet.hpp
